Move adding jquery to File Queue
