+++ src/Asset/File/Queue.php
@@ -5,6 +5,7 @@ use Bolt\Asset\AssetSortTrait;
 use Bolt\Asset\Injector;
 use Bolt\Asset\QueueInterface;
 use Bolt\Asset\Target;
+use Bolt\Config;
 use Bolt\Controller\Zone;
 use Symfony\Component\Asset\Packages;
 use Symfony\Component\HttpFoundation\Request;
@@ -24,6 +25,8 @@ class Queue implements QueueInterface
     protected $injector;
     /** @var Packages */
     protected $packages;
+    /** @var Config */
+    protected $config;
 
     /** @var FileAssetInterface[] */
     private $stylesheet = [];
@@ -35,11 +38,13 @@ class Queue implements QueueInterface
      *
      * @param Injector $injector
      * @param Packages $packages
+     * @param Config   $config
      */
-    public function __construct(Injector $injector, Packages $packages)
+    public function __construct(Injector $injector, Packages $packages, Config $config)
     {
         $this->injector = $injector;
         $this->packages = $packages;
+        $this->config = $config;
     }
 
     /**
@@ -75,6 +80,9 @@ class Queue implements QueueInterface
      */
     public function process(Request $request, Response $response)
     {
+        // Conditionally add jQuery
+        $this->addJquery($request, $response);
+
         /** @var FileAssetInterface $asset */
         foreach ($this->sort($this->javascript) as $key => $asset) {
             $this->processAsset($asset, $request, $response);
@@ -132,5 +140,43 @@ class Queue implements QueueInterface
         }
 
         $this->injector->inject($asset, $location, $response);
+    }
+
+    /**
+     * Insert jQuery, if it's not inserted already.
+     *
+     * Some of the patterns that 'match' are:
+     * - jquery.js
+     * - jquery.min.js
+     * - jquery-latest.js
+     * - jquery-latest.min.js
+     * - jquery-1.8.2.min.js
+     * - jquery-1.5.js
+     *
+     * @param Request  $request
+     * @param Response $response
+     */
+    protected function addJquery(Request $request, Response $response)
+    {
+        if (!$this->config->get('general/add_jquery', false) &&
+            !$this->config->get('theme/add_jquery', false)) {
+            return;
+        }
+
+        // Check zone to skip expensive regex
+        if (Zone::isFrontend($request) === false) {
+            return;
+        }
+
+        $html = $response->getContent();
+        $regex = '/<script(.*)jquery(-latest|-[0-9\.]*)?(\.min)?\.js/';
+        if (preg_match($regex, $html)) {
+            return;
+        }
+
+        $this->add(
+            (new JavaScript('js/jquery-2.2.4.min.js', 'bolt'))
+                ->setLocation(Target::BEFORE_JS)
+        );
     }
 }
